refactor for more scenarios
